Split enhancement tests
