+++ satpy/tests/enhancement_tests/test_enhancements.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2017-2023 Satpy developers
+# Copyright (c) 2017-2025 Satpy developers
 #
 # This file is part of satpy.
 #
@@ -15,540 +15,13 @@
 # satpy.  If not, see <http://www.gnu.org/licenses/>.
 """Unit testing the enhancements functions, e.g. cira_stretch."""
 
-import contextlib
 import os
-from tempfile import NamedTemporaryFile
 from unittest import mock
 
 import dask.array as da
 import numpy as np
 import pytest
 import xarray as xr
-
-from satpy.enhancements.color_mapping import create_colormap
-from satpy.enhancements.wrappers import on_dask_array, on_separate_bands, using_map_blocks
-
-# NOTE:
-# The following fixtures are not defined in this file, but are used and injected by Pytest:
-# - tmp_path
-
-
-def run_and_check_enhancement(func, data, expected, **kwargs):
-    """Perform basic checks that apply to multiple tests."""
-    pre_attrs = data.attrs
-    img = _get_enhanced_image(func, data, **kwargs)
-
-    _assert_image(img, pre_attrs, func.__name__, "palettes" in kwargs)
-    _assert_image_data(img, expected)
-
-
-def _get_enhanced_image(func, data, **kwargs):
-    from trollimage.xrimage import XRImage
-
-    img = XRImage(data)
-    func(img, **kwargs)
-
-    return img
-
-
-def _assert_image(img, pre_attrs, func_name, has_palette):
-    assert isinstance(img.data, xr.DataArray)
-    assert isinstance(img.data.data, da.Array)
-
-    old_keys = set(pre_attrs.keys())
-    # It is OK to have "enhancement_history" added
-    new_keys = set(img.data.attrs.keys()) - {"enhancement_history"}
-    # In case of palettes are used, _FillValue is added.
-    # Colorize doesn't add the fill value, so ignore that
-    if has_palette and func_name != "colorize":
-        assert "_FillValue" in new_keys
-        # Remove it from further comparisons
-        new_keys = new_keys - {"_FillValue"}
-    assert old_keys == new_keys
-
-
-def _assert_image_data(img, expected, dtype=None):
-    # Compute the data to mimic what xrimage geotiff writing does
-    res_data = img.data.data.compute()
-    assert not isinstance(res_data, da.Array)
-    np.testing.assert_allclose(res_data, expected, atol=1.e-6, rtol=0)
-    if dtype:
-        assert img.data.dtype == dtype
-        assert res_data.dtype == dtype
-
-
-def run_and_check_enhancement_with_dtype(func, data, expected, **kwargs):
-    """Perform basic checks that apply to multiple tests."""
-    pre_attrs = data.attrs
-    img = _get_enhanced_image(func, data, **kwargs)
-
-    _assert_image(img, pre_attrs, func.__name__, "palettes" in kwargs)
-    _assert_image_data(img, expected, dtype=data.dtype)
-
-
-def identical_decorator(func):
-    """Decorate but do nothing."""
-    return func
-
-
-class TestEnhancementStretch:
-    """Class for testing enhancements in satpy.enhancements."""
-
-    def setup_method(self):
-        """Create test data used by every test."""
-        data = np.arange(-210, 790, 100).reshape((2, 5)) * 0.95
-        data[0, 0] = np.nan  # one bad value for testing
-        crefl_data = np.arange(-210, 790, 100).reshape((2, 5)) * 0.95
-        crefl_data /= 5.605
-        crefl_data[0, 0] = np.nan  # one bad value for testing
-        crefl_data[0, 1] = 0.
-        self.ch1 = xr.DataArray(da.from_array(data, chunks=2), dims=("y", "x"), attrs={"test": "test"})
-        self.ch2 = xr.DataArray(da.from_array(crefl_data, chunks=2), dims=("y", "x"), attrs={"test": "test"})
-        rgb_data = np.stack([data, data, data])
-        self.rgb = xr.DataArray(da.from_array(rgb_data, chunks=(3, 2, 2)),
-                                dims=("bands", "y", "x"),
-                                coords={"bands": ["R", "G", "B"]})
-
-    @pytest.mark.parametrize(
-        ("decorator", "exp_call_cls"),
-        [
-            (identical_decorator, xr.DataArray),
-            (on_dask_array, da.Array),
-            (using_map_blocks, np.ndarray),
-        ],
-    )
-    @pytest.mark.parametrize("input_data_name", ["ch1", "ch2", "rgb"])
-    def test_enhancement_decorators(self, input_data_name, decorator, exp_call_cls):
-        """Test the utility decorators."""
-
-        def _enh_func(img):
-            def _calc_func(data):
-                assert isinstance(data, exp_call_cls)
-                return data
-
-            decorated_func = decorator(_calc_func)
-            return decorated_func(img.data)
-
-        in_data = getattr(self, input_data_name)
-        exp_data = in_data.values
-        if "bands" not in in_data.coords:
-            exp_data = exp_data[np.newaxis, :, :]
-        run_and_check_enhancement(_enh_func, in_data, exp_data)
-
-    @pytest.mark.parametrize("dtype", [np.float32, np.float64])
-    def test_cira_stretch(self, dtype):
-        """Test applying the cira_stretch."""
-        from satpy.enhancements.stretching import cira_stretch
-
-        expected = np.array([[
-            [np.nan, -7.04045974, -7.04045974, 0.79630132, 0.95947296],
-            [1.05181359, 1.11651012, 1.16635571, 1.20691137, 1.24110186]]], dtype=dtype)
-        run_and_check_enhancement_with_dtype(cira_stretch, self.ch1.astype(dtype), expected)
-
-    def test_reinhard(self):
-        """Test the reinhard algorithm."""
-        from satpy.enhancements.stretching import reinhard_to_srgb
-        expected = np.array([[[np.nan, 0., 0., 0.93333793, 1.29432402],
-                              [1.55428709, 1.76572249, 1.94738635, 2.10848544, 2.25432809]],
-
-                             [[np.nan, 0., 0., 0.93333793, 1.29432402],
-                              [1.55428709, 1.76572249, 1.94738635, 2.10848544, 2.25432809]],
-
-                             [[np.nan, 0., 0., 0.93333793, 1.29432402],
-                              [1.55428709, 1.76572249, 1.94738635, 2.10848544, 2.25432809]]])
-        run_and_check_enhancement(reinhard_to_srgb, self.rgb, expected)
-
-    def test_lookup(self):
-        """Test the lookup enhancement function."""
-        from satpy.enhancements.color_mapping import lookup
-        expected = np.array([[
-            [0., 0., 0., 0.333333, 0.705882],
-            [1., 1., 1., 1., 1.]]])
-        lut = np.arange(256.)
-        run_and_check_enhancement(lookup, self.ch1, expected, luts=lut)
-
-        expected = np.array([[[0., 0., 0., 0.333333, 0.705882],
-                              [1., 1., 1., 1., 1.]],
-                             [[0., 0., 0., 0.333333, 0.705882],
-                              [1., 1., 1., 1., 1.]],
-                             [[0., 0., 0., 0.333333, 0.705882],
-                              [1., 1., 1., 1., 1.]]])
-        lut = np.arange(256.)
-        lut = np.vstack((lut, lut, lut)).T
-        run_and_check_enhancement(lookup, self.rgb, expected, luts=lut)
-
-    def test_colorize(self):
-        """Test the colorize enhancement function."""
-        from trollimage.colormap import brbg
-
-        from satpy.enhancements.color_mapping import colorize
-        expected = np.array([
-            [[np.nan, 3.29411723e-01, 3.29411723e-01, 3.21825881e-08, 3.21825881e-08],
-             [3.21825881e-08, 3.21825881e-08, 3.21825881e-08, 3.21825881e-08, 3.21825881e-08]],
-            [[np.nan, 1.88235327e-01, 1.88235327e-01, 2.35294109e-01, 2.35294109e-01],
-             [2.35294109e-01, 2.35294109e-01, 2.35294109e-01, 2.35294109e-01, 2.35294109e-01]],
-            [[np.nan, 1.96078164e-02, 1.96078164e-02, 1.88235281e-01, 1.88235281e-01],
-             [1.88235281e-01, 1.88235281e-01, 1.88235281e-01, 1.88235281e-01, 1.88235281e-01]]])
-        run_and_check_enhancement(colorize, self.ch1, expected, palettes=brbg)
-
-    def test_palettize(self):
-        """Test the palettize enhancement function."""
-        from trollimage.colormap import brbg
-
-        from satpy.enhancements.color_mapping import palettize
-        expected = np.array([[[10, 0, 0, 10, 10], [10, 10, 10, 10, 10]]])
-        run_and_check_enhancement(palettize, self.ch1, expected, palettes=brbg)
-
-    def test_three_d_effect(self):
-        """Test the three_d_effect enhancement function."""
-        from satpy.enhancements.convolution import three_d_effect
-        expected = np.array([[
-            [np.nan, np.nan, -389.5, -294.5, 826.5],
-            [np.nan, np.nan, 85.5, 180.5, 1301.5]]])
-        run_and_check_enhancement(three_d_effect, self.ch1, expected)
-
-    def test_piecewise_linear_stretch(self):
-        """Test the piecewise_linear_stretch enhancement function."""
-        from satpy.enhancements.stretching import piecewise_linear_stretch
-        expected = np.array([[
-            [np.nan, 0., 0., 0.44378, 0.631734],
-            [0.737562, 0.825041, 0.912521, 1., 1.]]])
-        run_and_check_enhancement(piecewise_linear_stretch,
-                                  self.ch2 / 100.0,
-                                  expected,
-                                  xp=[0., 25., 55., 100., 255.],
-                                  fp=[0., 90., 140., 175., 255.],
-                                  reference_scale_factor=255,
-                                  )
-
-    def test_btemp_threshold(self):
-        """Test applying the cira_stretch."""
-        from satpy.enhancements.stretching import btemp_threshold
-
-        expected = np.array([[
-            [np.nan, 0.946207, 0.892695, 0.839184, 0.785672],
-            [0.73216, 0.595869, 0.158745, -0.278379, -0.715503]]])
-        run_and_check_enhancement(btemp_threshold, self.ch1, expected,
-                                  min_in=-200, max_in=500, threshold=350)
-
-    def test_merge_colormaps(self):
-        """Test merging colormaps."""
-        from trollimage.colormap import Colormap
-
-        from satpy.enhancements.color_mapping import _merge_colormaps as mcp
-        from satpy.enhancements.color_mapping import create_colormap
-        ret_map = mock.MagicMock()
-
-        create_colormap_mock = mock.Mock(wraps=create_colormap)
-        cmap1 = Colormap((1, (1., 1., 1.)))
-        kwargs = {"palettes": cmap1}
-
-        with mock.patch("satpy.enhancements.color_mapping.create_colormap", create_colormap_mock):
-            res = mcp(kwargs)
-        assert res is cmap1
-        create_colormap_mock.assert_not_called()
-        create_colormap_mock.reset_mock()
-        ret_map.reset_mock()
-
-        cmap1 = {"colors": "blues", "min_value": 0,
-                 "max_value": 1}
-        kwargs = {"palettes": [cmap1]}
-        with mock.patch("satpy.enhancements.color_mapping.create_colormap", create_colormap_mock), \
-                mock.patch("trollimage.colormap.blues", ret_map):
-            _ = mcp(kwargs)
-        create_colormap_mock.assert_called_once()
-        ret_map.reverse.assert_not_called()
-        ret_map.set_range.assert_called_with(0, 1)
-        create_colormap_mock.reset_mock()
-        ret_map.reset_mock()
-
-        cmap2 = {"colors": "blues", "min_value": 2,
-                 "max_value": 3, "reverse": True}
-        kwargs = {"palettes": [cmap2]}
-        with mock.patch("trollimage.colormap.blues", ret_map):
-            _ = mcp(kwargs)
-        ret_map.reverse.assert_called_once()
-        ret_map.set_range.assert_called_with(2, 3)
-        create_colormap_mock.reset_mock()
-        ret_map.reset_mock()
-
-        kwargs = {"palettes": [cmap1, cmap2]}
-        with mock.patch("trollimage.colormap.blues", ret_map):
-            _ = mcp(kwargs)
-        ret_map.__add__.assert_called_once()
-
-    def tearDown(self):
-        """Clean up."""
-
-
-@contextlib.contextmanager
-def closed_named_temp_file(**kwargs):
-    """Named temporary file context manager that closes the file after creation.
-
-    This helps with Windows systems which can get upset with opening or
-    deleting a file that is already open.
-
-    """
-    try:
-        with NamedTemporaryFile(delete=False, **kwargs) as tmp_cmap:
-            yield tmp_cmap.name
-    finally:
-        os.remove(tmp_cmap.name)
-
-
-def _write_cmap_to_file(cmap_filename, cmap_data):
-    ext = os.path.splitext(cmap_filename)[1]
-    if ext in (".npy",):
-        np.save(cmap_filename, cmap_data)
-    elif ext in (".npz",):
-        np.savez(cmap_filename, cmap_data)
-    else:
-        np.savetxt(cmap_filename, cmap_data, delimiter=",")
-
-
-def _generate_cmap_test_data(color_scale, colormap_mode):
-    cmap_data = np.array([
-        [1, 0, 0],
-        [1, 1, 0],
-        [1, 1, 1],
-        [0, 0, 1],
-    ], dtype=np.float64)
-    if len(colormap_mode) != 3:
-        _cmap_data = cmap_data
-        cmap_data = np.empty((cmap_data.shape[0], len(colormap_mode)),
-                             dtype=np.float64)
-        if colormap_mode.startswith("V") or colormap_mode.endswith("A"):
-            cmap_data[:, 0] = np.array([128, 130, 132, 134]) / 255.0
-            cmap_data[:, -3:] = _cmap_data
-        if colormap_mode.startswith("V") and colormap_mode.endswith("A"):
-            cmap_data[:, 1] = np.array([128, 130, 132, 134]) / 255.0
-    if color_scale is None or color_scale == 255:
-        cmap_data = (cmap_data * 255).astype(np.uint8)
-    return cmap_data
-
-
-class TestColormapLoading:
-    """Test utilities used with colormaps."""
-
-    @pytest.mark.parametrize("color_scale", [None, 1.0])
-    @pytest.mark.parametrize("colormap_mode", ["RGB", "VRGB", "VRGBA"])
-    @pytest.mark.parametrize("extra_kwargs",
-                             [
-                                 {},
-                                 {"min_value": 50, "max_value": 100},
-                             ])
-    @pytest.mark.parametrize("filename_suffix", [".npy", ".npz", ".csv"])
-    def test_cmap_from_file(self, color_scale, colormap_mode, extra_kwargs, filename_suffix):
-        """Test that colormaps can be loaded from a binary file."""
-        # create the colormap file on disk
-        with closed_named_temp_file(suffix=filename_suffix) as cmap_filename:
-            cmap_data = _generate_cmap_test_data(color_scale, colormap_mode)
-            _write_cmap_to_file(cmap_filename, cmap_data)
-
-            unset_first_value = 128.0 / 255.0 if colormap_mode.startswith("V") else 0.0
-            unset_last_value = 134.0 / 255.0 if colormap_mode.startswith("V") else 1.0
-            if (color_scale is None or color_scale == 255) and colormap_mode.startswith("V"):
-                unset_first_value *= 255
-                unset_last_value *= 255
-            if "min_value" in extra_kwargs:
-                unset_first_value = extra_kwargs["min_value"]
-                unset_last_value = extra_kwargs["max_value"]
-
-            first_color = [1.0, 0.0, 0.0]
-            if colormap_mode == "VRGBA":
-                first_color = [128.0 / 255.0] + first_color
-
-            kwargs1 = {"filename": cmap_filename}
-            kwargs1.update(extra_kwargs)
-            if color_scale is not None:
-                kwargs1["color_scale"] = color_scale
-
-            cmap = create_colormap(kwargs1)
-
-        assert cmap.colors.shape[0] == 4
-        np.testing.assert_equal(cmap.colors[0], first_color)
-        assert cmap.values.shape[0] == 4
-        assert cmap.values[0] == unset_first_value
-        assert cmap.values[-1] == unset_last_value
-
-    def test_cmap_vrgb_as_rgba(self):
-        """Test that data created as VRGB still reads as RGBA."""
-        with closed_named_temp_file(suffix=".npy") as cmap_filename:
-            cmap_data = _generate_cmap_test_data(None, "VRGB")
-            np.save(cmap_filename, cmap_data)
-            cmap = create_colormap({"filename": cmap_filename, "colormap_mode": "RGBA"})
-
-        assert cmap.colors.shape[0] == 4
-        assert cmap.colors.shape[1] == 4  # RGBA
-        np.testing.assert_equal(cmap.colors[0], [128 / 255., 1.0, 0, 0])
-        assert cmap.values.shape[0] == 4
-        assert cmap.values[0] == 0
-        assert cmap.values[-1] == 1.0
-
-    def test_cmap_with_alpha_set(self):
-        """Test that the min_alpha and max_alpha arguments set the alpha channel correctly."""
-        with closed_named_temp_file(suffix=".npy") as cmap_filename:
-            cmap_data = _generate_cmap_test_data(None, "RGB")
-            np.save(cmap_filename, cmap_data)
-            cmap = create_colormap({"filename": cmap_filename, "min_alpha": 100, "max_alpha": 255})
-
-        assert cmap.colors.shape[0] == 4
-        assert cmap.colors.shape[1] == 4  # RGBA
-        # check that we start from min_alpha
-        np.testing.assert_equal(cmap.colors[0], [1.0, 0, 0, 100/255.])
-        # two thirds of the linear scale
-        np.testing.assert_almost_equal(cmap.colors[2], [1., 1., 1., (100+(2/3)*(255-100))/255])
-        # check that we end at max_alpha
-        np.testing.assert_equal(cmap.colors[3], [0, 0, 1., 1.0])
-        # check that values have not been changed
-        assert cmap.values.shape[0] == 4
-        assert cmap.values[0] == 0
-        assert cmap.values[-1] == 1.0
-
-    @pytest.mark.parametrize("alpha_arg", ["max_alpha", "min_alpha"])
-    def test_cmap_error_with_only_one_alpha_set(self, alpha_arg):
-        """Test that when only min_alpha or max_alpha arguments are set an error is raised."""
-        with closed_named_temp_file(suffix=".npy") as cmap_filename:
-            cmap_data = _generate_cmap_test_data(None, "RGB")
-            np.save(cmap_filename, cmap_data)
-
-            # check that if a value is missing we raise a ValueError
-            with pytest.raises(ValueError, match="Both 'min_alpha' and 'max_alpha' must be specified*."):
-                create_colormap({"filename": cmap_filename, alpha_arg: 255})
-
-    @pytest.mark.parametrize(
-        ("real_mode", "forced_mode"),
-        [
-            ("VRGBA", "RGBA"),
-            ("VRGBA", "VRGB"),
-            ("RGBA", "RGB"),
-        ]
-    )
-    @pytest.mark.parametrize("filename_suffix", [".npy", ".csv"])
-    def test_cmap_bad_mode(self, real_mode, forced_mode, filename_suffix):
-        """Test that reading colormaps with the wrong mode fails."""
-        with closed_named_temp_file(suffix=filename_suffix) as cmap_filename:
-            cmap_data = _generate_cmap_test_data(None, real_mode)
-            _write_cmap_to_file(cmap_filename, cmap_data)
-            # Force colormap_mode VRGBA to RGBA and we should see an exception
-            with pytest.raises(ValueError, match="Unexpected colormap shape for mode .*"):
-                create_colormap({"filename": cmap_filename, "colormap_mode": forced_mode})
-
-    def test_cmap_from_file_bad_shape(self):
-        """Test that unknown array shape causes an error."""
-        from satpy.enhancements.color_mapping import create_colormap
-
-        # create the colormap file on disk
-        with closed_named_temp_file(suffix=".npy") as cmap_filename:
-            np.save(cmap_filename, np.array([
-                [0],
-                [64],
-                [128],
-                [255],
-            ]))
-
-            with pytest.raises(ValueError, match="Unexpected colormap shape for mode 'None'"):
-                create_colormap({"filename": cmap_filename})
-
-    def test_cmap_from_config_path(self, tmp_path):
-        """Test loading a colormap relative to a config path."""
-        import satpy
-        from satpy.enhancements.color_mapping import create_colormap
-
-        cmap_dir = tmp_path / "colormaps"
-        cmap_dir.mkdir()
-        cmap_filename = cmap_dir / "my_colormap.npy"
-        cmap_data = _generate_cmap_test_data(None, "RGBA")
-        np.save(cmap_filename, cmap_data)
-        with satpy.config.set(config_path=[tmp_path]):
-            rel_cmap_filename = os.path.join("colormaps", "my_colormap.npy")
-            cmap = create_colormap({"filename": rel_cmap_filename, "colormap_mode": "RGBA"})
-
-        assert cmap.colors.shape[0] == 4
-        assert cmap.colors.shape[1] == 4  # RGBA
-        np.testing.assert_equal(cmap.colors[0], [128 / 255., 1.0, 0, 0])
-        assert cmap.values.shape[0] == 4
-        assert cmap.values[0] == 0
-        assert cmap.values[-1] == 1.0
-
-    def test_cmap_from_trollimage(self):
-        """Test that colormaps in trollimage can be loaded."""
-        from satpy.enhancements.color_mapping import create_colormap
-        cmap = create_colormap({"colors": "pubu"})
-        from trollimage.colormap import pubu
-        np.testing.assert_equal(cmap.colors, pubu.colors)
-        np.testing.assert_equal(cmap.values, pubu.values)
-
-    def test_cmap_no_colormap(self):
-        """Test that being unable to create a colormap raises an error."""
-        from satpy.enhancements.color_mapping import create_colormap
-        with pytest.raises(ValueError, match="Unknown colormap format: .*"):
-            create_colormap({})
-
-    def test_cmap_list(self):
-        """Test that colors can be a list/tuple."""
-        from satpy.enhancements.color_mapping import create_colormap
-        colors = [
-            [0., 0., 1.],
-            [1., 0., 1.],
-            [0., 1., 1.],
-            [1., 1., 1.],
-        ]
-        values = [2, 4, 6, 8]
-        cmap = create_colormap({"colors": colors, "color_scale": 1})
-        assert cmap.colors.shape[0] == 4
-        np.testing.assert_equal(cmap.colors[0], [0.0, 0.0, 1.0])
-        assert cmap.values.shape[0] == 4
-        assert cmap.values[0] == 0
-        assert cmap.values[-1] == 1.0
-
-        cmap = create_colormap({"colors": colors, "color_scale": 1, "values": values})
-        assert cmap.colors.shape[0] == 4
-        np.testing.assert_equal(cmap.colors[0], [0.0, 0.0, 1.0])
-        assert cmap.values.shape[0] == 4
-        assert cmap.values[0] == 2
-        assert cmap.values[-1] == 8
-
-
-def test_on_separate_bands():
-    """Test the `on_separate_bands` decorator."""
-
-    def func(array, index, gain=2):
-        return xr.DataArray(np.ones(array.shape, dtype=array.dtype) * index * gain,
-                            coords=array.coords, dims=array.dims, attrs=array.attrs)
-
-    separate_func = on_separate_bands(func)
-    arr = xr.DataArray(np.zeros((3, 10, 10)), dims=["bands", "y", "x"], coords={"bands": ["R", "G", "B"]})
-    assert separate_func(arr).shape == arr.shape
-    assert all(separate_func(arr, gain=1).values[:, 0, 0] == [0, 1, 2])
-
-
-def test_using_map_blocks():
-    """Test the `using_map_blocks` decorator."""
-
-    def func(np_array, block_info=None):
-        value = block_info[0]["chunk-location"][-1]
-        return np.ones(np_array.shape) * value
-
-    map_blocked_func = using_map_blocks(func)
-    arr = xr.DataArray(da.zeros((3, 10, 10), dtype=int, chunks=5), dims=["bands", "y", "x"])
-    res = map_blocked_func(arr)
-    assert res.shape == arr.shape
-    assert res[0, 0, 0].compute() != res[0, 9, 9].compute()
-
-
-def test_on_dask_array():
-    """Test the `on_dask_array` decorator."""
-
-    def func(dask_array):
-        if not isinstance(dask_array, da.core.Array):
-            pytest.fail("Array is not a dask array")
-        return dask_array
-
-    dask_func = on_dask_array(func)
-    arr = xr.DataArray(da.zeros((3, 10, 10), dtype=int, chunks=5), dims=["bands", "y", "x"])
-    res = dask_func(arr)
-    assert res.shape == arr.shape
 
 
 @pytest.fixture
@@ -702,45 +175,3 @@ def test_nwcsaf_comps(fake_area, tmp_path, data):
                 ((fake_alti - rng[0]) * (255 / np.ptp(rng))).round())
         else:
             np.testing.assert_allclose(im.data.sel(bands="P"), fake_alti)
-
-
-class TestTCREnhancement:
-    """Test the AHI enhancement functions."""
-
-    def setup_method(self):
-        """Create test data."""
-        data = da.arange(-100, 1000, 110).reshape(2, 5)
-        rgb_data = np.stack([data, data, data])
-        self.rgb = xr.DataArray(rgb_data, dims=("bands", "y", "x"),
-                                coords={"bands": ["R", "G", "B"]},
-                                attrs={"platform_name": "Himawari-8"})
-
-    def test_jma_true_color_reproduction(self):
-        """Test the jma_true_color_reproduction enhancement."""
-        from trollimage.xrimage import XRImage
-
-        from satpy.enhancements.stretching import jma_true_color_reproduction
-
-        expected = [[[-109.93, 10.993, 131.916, 252.839, 373.762],
-                     [494.685, 615.608, 736.531, 857.454, 978.377]],
-
-                    [[-97.73, 9.773, 117.276, 224.779, 332.282],
-                     [439.785, 547.288, 654.791, 762.294, 869.797]],
-
-                    [[-93.29, 9.329, 111.948, 214.567, 317.186],
-                     [419.805, 522.424, 625.043, 727.662, 830.281]]]
-
-        img = XRImage(self.rgb)
-        jma_true_color_reproduction(img)
-
-        np.testing.assert_almost_equal(img.data.compute(), expected)
-
-        self.rgb.attrs["platform_name"] = None
-        img = XRImage(self.rgb)
-        with pytest.raises(ValueError, match="Missing platform name."):
-            jma_true_color_reproduction(img)
-
-        self.rgb.attrs["platform_name"] = "Fakesat"
-        img = XRImage(self.rgb)
-        with pytest.raises(KeyError, match="No conversion matrix found for platform Fakesat"):
-            jma_true_color_reproduction(img)
